requirements.txt
python-telegram-bot==20.3
google-generativeai>=0.3.0
python-dotenv>=0.19.0